local_db/requirements.txt
alpaca-py==0.21.0
annotated-types==0.6.0
azure-core==1.30.1
azure-cosmos==4.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
msgpack==1.0.8
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sseclient-py==1.8.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
wheel==0.43.0
